Split out reader
