+++ src/side-panel.html
@@ -18,8 +18,8 @@
 					data-i18n-title="showPageVariables">
 					<i data-lucide="ellipsis"></i>
 				</a>
-				<a href="#" id="tidy-mode" class="clickable-icon" 
-					data-i18n-title="tidyPage">
+				<a href="#" id="reader-mode" class="clickable-icon" 
+					data-i18n-title="reader">
 					<i data-lucide="book-open"></i>
 				</a>
 				<a href="#" id="highlighter-mode" class="clickable-icon"
